+++ src/test/java/com/FinWorks/runners/CukesRunner.java
@@ -13,7 +13,7 @@ import org.junit.runner.RunWith;
         features = "src/test/resources/features",
         glue = "com/FinWorks/step_definitions",
         dryRun = false,
-        tags = "@wip"
+        tags = "@login"
 )
 public class CukesRunner {
 }
