+++ src/test/java/com/FinWorks/runners/EvelinCukesRunner.java
@@ -1,0 +1,4 @@
+package com.FinWorks.runners;
+
+public class EvelinCukesRunner {
+}
